Move the other common components into new folder
Also changed the pages to use absolute paths.
