+++ source/components/common/layout.tsx
@@ -1,4 +1,4 @@
-import Navbar from 'components/navbar';
+import Navbar from 'components/common/navbar';
 
 export const Layout = (): JSX.Element => (
   <>
